+++ lib/dynomite/errors.rb
@@ -1,0 +1,15 @@
+module Dynomite
+  module Errors
+    class ValidationError < StandardError
+      def initialize(msg)
+        super
+      end
+    end
+
+    class ReservedWordError < StandardError
+      def initialize(msg)
+        super
+      end
+    end
+  end
+end